api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
click==8.1.7
coloredlogs==15.0.1
fastapi==0.110.0
flatbuffers==24.3.25
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
multidict==6.0.5
numpy==1.26.4
onnxruntime==1.17.1
orjson==3.9.15
packaging==24.0
pillow==10.2.0
protobuf==5.26.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
